Extract agreement helper
CreateAgreement and ViewAgreements create similar responses so extracting this helper will allow us to share this implementation between them
