+++ lib/hackney/income/use_case_factory.rb
@@ -198,11 +198,11 @@ module Hackney
       end
 
       def view_agreements
-        Hackney::Income::ViewAgreements
+        Hackney::Income::ViewAgreements.new
       end
 
       def create_agreement
-        Hackney::Income::CreateAgreement
+        Hackney::Income::CreateAgreement.new
       end
 
       private
